Fix success status code being written to file or mishandled for download case
When downloading a file, we would have issues in which the server's status code could accidentally be included in the file buffer. Because we're dealing with data streams, there's no direct way to separate the data. It would've been possible to use a delimeter, but here we're just sending the file's length beforehand, and then only reading exactly that many bytes into the file -- thereby separating the file content to the status code.